feat seo
add seo stuff and better description
reshape some methods
